+++ packages/atom.io/internal/src/subscribe/index.ts
@@ -1,3 +1,5 @@
+export * from "./recall-state"
+export * from "./subscribe-in-store"
 export * from "./subscribe-to-root-atoms"
 export * from "./subscribe-to-state"
 export * from "./subscribe-to-timeline"
