+++ packages/atom.io/src/subscribe.ts
@@ -1,11 +1,5 @@
-import type { Flat, Func, Store } from "atom.io/internal"
-import {
-	arbitrary,
-	IMPLICIT,
-	subscribeToState,
-	subscribeToTimeline,
-	subscribeToTransaction,
-} from "atom.io/internal"
+import type { Flat, Func } from "atom.io/internal"
+import { arbitrary, IMPLICIT, subscribeInStore } from "atom.io/internal"
 
 import type {
 	FamilyMetadata,
@@ -30,52 +24,6 @@ export type UpdateHandler<T> = (update: StateUpdate<T>) => void
 export type TransactionUpdateHandler<F extends Func> = (
 	data: TransactionUpdate<F>,
 ) => void
-
-export function subscribeInStore<T>(
-	store: Store,
-	token: ReadableToken<T>,
-	handleUpdate: UpdateHandler<T>,
-	key?: string,
-): () => void
-export function subscribeInStore<F extends Func>(
-	store: Store,
-	token: TransactionToken<F>,
-	handleUpdate: TransactionUpdateHandler<F>,
-	key?: string,
-): () => void
-export function subscribeInStore<M extends TimelineManageable>(
-	store: Store,
-	token: TimelineToken<M>,
-	handleUpdate: (update: TimelineUpdate<M> | `redo` | `undo`) => void,
-	key?: string,
-): () => void
-export function subscribeInStore<M extends TimelineManageable>(
-	store: Store,
-	token: ReadableToken<any> | TimelineToken<M> | TransactionToken<any>,
-	handleUpdate:
-		| TransactionUpdateHandler<any>
-		| UpdateHandler<any>
-		| ((update: TimelineUpdate<M> | `redo` | `undo`) => void),
-	key?: string,
-): () => void
-export function subscribeInStore(
-	store: Store,
-	token: ReadableToken<any> | TimelineToken<any> | TransactionToken<any>,
-	handleUpdate: (update: any) => void,
-	key: string = arbitrary(),
-): () => void {
-	switch (token.type) {
-		case `atom`:
-		case `mutable_atom`:
-		case `readonly_selector`:
-		case `selector`:
-			return subscribeToState(token, handleUpdate, key, store)
-		case `transaction`:
-			return subscribeToTransaction(token, handleUpdate, key, store)
-		case `timeline`:
-			return subscribeToTimeline(token, handleUpdate, key, store)
-	}
-}
 
 export function subscribe<T>(
 	token: ReadableToken<T>,
